% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{pubmed_search}
\alias{pubmed_search}
\title{search pubmed via rentrez}
\usage{
pubmed_search(term, retmax = "1000000", verbose = F)
}
\arguments{
\item{term}{search term for pubmed}

\item{retmax}{integer Maximum number of hits returned by the search (passed down)}

\item{verbose}{logical, if T will display request and response via rentrez::entrez_search}
}
\value{
a character vector of one or more ids matching search or NA
}
\description{
this wrapper functions returns a tibble of the key variables
for each hit.
}
\examples{
  # pubmed_search("GSE8305")
}

