% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{pubmed_summary}
\alias{pubmed_summary}
\title{get summary tibble matching one or more pubmed ids}
\usage{
pubmed_summary(ids)
}
\arguments{
\item{ids}{character vector of pubmed db ids, such as returned from pubmed_search()}
}
\value{
a tibble with uid, sortpubdate, elocationid, sortfirstauthor, lastauthot,
 source, and sorttitle vars of length(ids) observations.
}
\description{
get summary tibble matching one or more pubmed ids
}

