% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_pubmed}
\alias{search_pubmed}
\title{search pubmed via rentrez}
\usage{
search_pubmed(term, retmax = 20)
}
\arguments{
\item{term}{search term for pubmed}

\item{retmax}{integer Maximum number of hits returned by the search (passed down)}
}
\description{
this wrapper functions returns a tibble of the key variables
for each hit.
}
\examples{
  # search_pubmed("GSE8305")
}

