% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{to_tibble}
\alias{to_tibble}
\alias{to_tibble.default}
\alias{to_tibble.esummary}
\alias{to_tibble.esummary_list}
\title{convert esummary like object into tibble}
\usage{
to_tibble(x)

\method{to_tibble}{esummary}(x)

\method{to_tibble}{esummary_list}(x)

\method{to_tibble}{default}(x)
}
\arguments{
\item{x}{esummary object}
}
\description{
convert esummary like object into tibble
}

